refactor: turn repo into a monorepo
