+++ packages/qwik-router/package.json
@@ -1,0 +1,71 @@
+{
+  "name": "qwik-router",
+  "version": "0.5.0",
+  "description": "A SPA-like router for Qwik.",
+  "license": "MIT",
+  "private": false,
+  "author": {
+    "name": "Danny França",
+    "email": "me@dannyfranca.com"
+  },
+  "contributors": [
+    {
+      "name": "Danny França <me@dannyfranca.com>"
+    }
+  ],
+  "repository": "https://github.com/dannyfranca/qwik-router",
+  "publishConfig": {
+    "access": "public"
+  },
+  "keywords": [
+    "qwik",
+    "router",
+    "spa",
+    "single-page-application"
+  ],
+  "type": "module",
+  "main": "./lib/index.qwik.mjs",
+  "qwik": "./lib/index.qwik.mjs",
+  "types": "./lib-types/index.d.ts",
+  "exports": {
+    ".": {
+      "import": "./lib/index.qwik.mjs",
+      "require": "./lib/index.qwik.cjs",
+      "types": "./lib-types/index.d.ts"
+    }
+  },
+  "files": [
+    "lib",
+    "lib-types"
+  ],
+  "engines": {
+    "node": ">=15.0.0"
+  },
+  "scripts": {
+    "build": "qwik build",
+    "build.lib": "vite build --mode lib",
+    "build.types": "tsc --emitDeclarationOnly",
+    "dev": "vite --mode ssr",
+    "dev.debug": "node --inspect-brk ./node_modules/vite/bin/vite.js --mode ssr --force",
+    "test": "vitest",
+    "fmt": "prettier --write .",
+    "fmt.check": "prettier --check .",
+    "lint": "eslint \"src/**/*.ts*\"",
+    "release": "np",
+    "start": "vite --open --mode ssr",
+    "qwik": "qwik"
+  },
+  "devDependencies": {
+    "@vitest/coverage-istanbul": "0.33.0",
+    "eslint-plugin-qwik": "latest",
+    "happy-dom": "10.0.3",
+    "path-to-regexp": "6.2.1",
+    "undici": "5.22.1",
+    "vite": "4.4.2",
+    "vite-tsconfig-paths": "4.2.0",
+    "vitest": "0.33.0"
+  },
+  "peerDependencies": {
+    "@builder.io/qwik": ">=0.24.0"
+  }
+}
