+++ packages/qwik-router/src/types.ts
