+++ packages/qwik-router/src/utils/generate-new-url-from-path.ts
