+++ packages/qwik-router/src/utils/get-window.test.ts
@@ -1,4 +1,4 @@
-import { getWindow } from "@/utils/get-window";
+import { getWindow } from "packages/qwik-router/src/utils/get-window";
 
 describe(getWindow.name, () => {
   it("Does not return the window", async () => {
