+++ packages/qwik-router/src/utils/get-window.ts
