+++ packages/qwik-router/src/utils/update-route-state.ts
