+++ packages/qwik-router/tsconfig.json
@@ -1,21 +1,11 @@
 {
+  "extends": "../../tsconfig.base.json",
   "compilerOptions": {
     "allowJs": true,
     "target": "ES2017",
     "module": "ES2020",
-    "lib": ["es2020", "DOM"],
     "jsx": "react-jsx",
     "jsxImportSource": "@builder.io/qwik",
-    "strict": true,
-    "declaration": true,
-    "declarationDir": "lib-types",
-    "resolveJsonModule": true,
-    "moduleResolution": "node",
-    "esModuleInterop": true,
-    "skipLibCheck": true,
-    "incremental": true,
-    "isolatedModules": true,
-    "importHelpers": true,
     "types": ["vitest/globals", "vite/client"],
     "paths": {
       "@/*": ["./src/*"],
